Add creat new pancake when pouring into a pan
a pancake is created when a batter ball (the invisbal physics spher w/ trail render) comes into contact with a frying pan.
Each batter ball containes the amount of batter poured across the batter ball spwan intervals.
The Pan adds the batter to the new pancake.
batter can only be added to a pancake if it is raw.
when batter is added to a pancake it grows in hight and spreads out over time to simulate mass.

Any pancakes that are all ready in the pan will become part of the new pancake.