Added pancake cooking
So the pancakes have 2 sides :) so the way this works is both sides start as batter (derr) once batter is compleat both sides go to a raw state and update iderpendently of one another.

the cooking time is manipulated by the pan temp.
iv set up the state in the prefab
color of pancake updates depending on the state

TODO:
still need to detect which side is cooking.